+++ opendaylight/md-sal/samples/clustering-test-app/provider/src/main/java/org/opendaylight/controller/clustering/it/provider/impl/AbstractTransactionHandler.java
@@ -81,8 +81,8 @@ abstract class AbstractTransactionHandler {
     private volatile State state;
 
     AbstractTransactionHandler(final TransactionsParams params) {
-        runtimeNanos = TimeUnit.SECONDS.toNanos(params.getSeconds());
-        delayNanos = SECOND_AS_NANO / params.getTransactionsPerSecond();
+        runtimeNanos = TimeUnit.SECONDS.toNanos(params.getSeconds().longValue());
+        delayNanos = SECOND_AS_NANO / params.getTransactionsPerSecond().longValue();
     }
 
     final synchronized void doStart() {
